setup DungeonWorldController
