+++ core/src/de/fhbielefeld/pmdungeon/vorgaben/game/DungeonGame.java
@@ -1,0 +1,25 @@
+package de.fhbielefeld.pmdungeon.vorgaben.game;
+import com.badlogic.gdx.Game;
+import com.badlogic.gdx.graphics.g2d.SpriteBatch;
+import de.fhbielefeld.pmdungeon.vorgaben.game.Controller.MainGameController;
+
+
+public class DungeonGame extends Game {
+
+    private SpriteBatch batch;
+
+    @Override
+    public void create() {
+        batch = new SpriteBatch();
+        this.setScreen(new MainGameController(this));
+    }
+
+    @Override
+    public void dispose() {
+        batch.dispose();
+    }
+
+    public SpriteBatch getBatch() {
+        return batch;
+    }
+}
