--- core/src/de/fhbielefeld/pmdungeon/vorgaben/game/GameScreen.java
@@ -1,95 +1,0 @@
-package de.fhbielefeld.pmdungeon.vorgaben.game;
-
-import com.badlogic.gdx.Gdx;
-import com.badlogic.gdx.Input;
-import com.badlogic.gdx.ScreenAdapter;
-import com.badlogic.gdx.graphics.OrthographicCamera;
-
-import static com.badlogic.gdx.graphics.GL20.GL_COLOR_BUFFER_BIT;
-
-/**
- * Class for setting up the basics like window height and camera. Is also holding the gameworld instance and is handling the demo sequence.
- */
-public class GameScreen extends ScreenAdapter {
-
-    public static final float VIRTUAL_HEIGHT = 5f;
-
-    private final PMDungeon pmDungeon;
-    private OrthographicCamera camera;
-
-    private final GameController gameController;
-    private final DemoSequence demoSequence;
-
-    public GameScreen(final PMDungeon pmDungeon) {
-        this.pmDungeon = pmDungeon;
-        this.gameController = new GameController(pmDungeon.getBatch());
-       // this.hud = new HeadUpDisplay(gameWorld);
-        this.demoSequence = new DemoSequence(gameController);
-        setupCamera();
-    }
-
-    /**
-     * Setting up the camera.
-     */
-    private void setupCamera() {
-        camera = new OrthographicCamera();
-        camera.position.set(0, 0, 0);
-        camera.zoom +=3;
-        camera.update();
-    }
-
-    /**
-     * Variable camera zoom for developing purposes.
-     */
-    private void debugCameraZoom() {
-        float cameraZoomSpeed = 5;
-        if (Gdx.input.isKeyPressed(Input.Keys.PLUS)) {
-            camera.zoom -= cameraZoomSpeed * Gdx.graphics.getDeltaTime();
-        }
-        if (Gdx.input.isKeyPressed(Input.Keys.MINUS)) {
-            camera.zoom += cameraZoomSpeed * Gdx.graphics.getDeltaTime();
-        }
-    }
-
-    /**
-     * Main gameloop.
-     *
-     * @param delta Time since last loop.
-     */
-    @Override
-    public void render(float delta) {
-        Gdx.gl.glClearColor(0, 0, 0, 1);
-        Gdx.gl.glClear(GL_COLOR_BUFFER_BIT);
-
-        //debugCameraZoom();
-
-        demoSequence.update();
-        gameController.update();
-
-       // camera.position.set(gameWorld.getHero().getPositionX(), gameWorld.getHero().getPositionY(), 0);
-        camera.position.set(0,0,0);
-        camera.update();
-        pmDungeon.getBatch().setProjectionMatrix(camera.combined);
-
-        pmDungeon.getBatch().begin();
-        gameController.render();
-        pmDungeon.getBatch().end();
-    }
-
-    /**
-     * Resizing the camera according to the size of the window.
-     *
-     * @param width  Window width
-     * @param height Window height
-     */
-    @Override
-    public void resize(int width, int height) {
-        camera.setToOrtho(false, VIRTUAL_HEIGHT * width / (float) height, VIRTUAL_HEIGHT);
-        pmDungeon.getBatch().setProjectionMatrix(camera.combined);
-    }
-
-    @Override
-    public void dispose() {
-        gameController.dispose();
-    }
-}
