--- core/src/de/fhbielefeld/pmdungeon/vorgaben/game/PMDungeon.java
@@ -1,24 +1,0 @@
-package de.fhbielefeld.pmdungeon.vorgaben.game;
-import com.badlogic.gdx.Game;
-import com.badlogic.gdx.graphics.g2d.SpriteBatch;
-
-
-public class PMDungeon extends Game {
-
-    private SpriteBatch batch;
-
-    @Override
-    public void create() {
-        batch = new SpriteBatch();
-        this.setScreen(new GameScreen(this));
-    }
-
-    @Override
-    public void dispose() {
-        batch.dispose();
-    }
-
-    public SpriteBatch getBatch() {
-        return batch;
-    }
-}
