+++ core/src/de/fhbielefeld/pmdungeon/vorgaben/interfaces/IUpdateable.java
@@ -1,0 +1,14 @@
+package de.fhbielefeld.pmdungeon.vorgaben.interfaces;
+
+public interface IUpdateable {
+
+    /**
+     * @return if this instance can be deletet (removed from entitycontroll);
+     */
+    public boolean deleteable();
+
+    /**
+     * Will be executed every frame. Remember to draw/animate your drawable objects.
+     */
+    public void update();
+}
