+++ desktop/src/de/fhbielefeld/pmdungeon/desktop/DesktopLauncher.java
@@ -2,7 +2,7 @@ package de.fhbielefeld.pmdungeon.desktop;
 
 import com.badlogic.gdx.backends.lwjgl3.Lwjgl3Application;
 import com.badlogic.gdx.backends.lwjgl3.Lwjgl3ApplicationConfiguration;
-import de.fhbielefeld.pmdungeon.vorgaben.game.PMDungeon;
+import de.fhbielefeld.pmdungeon.vorgaben.game.DungeonGame;
 
 
 public class DesktopLauncher {
@@ -10,6 +10,6 @@ public class DesktopLauncher {
 		Lwjgl3ApplicationConfiguration config = new Lwjgl3ApplicationConfiguration();
 		config.setTitle("PM-Dungeon");
         config.setForegroundFPS(30);
-		new Lwjgl3Application(new PMDungeon(), config);
+		new Lwjgl3Application(new DungeonGame(), config);
 	}
 }
